Everything works, pretty UI added

+++ Assets/Game/Scripts/Common/IntValueView.cs
@@ -1,4 +1,4 @@
 
-public class IntValueView : ValueView<int>
+public class IntValueView : ValueView<float>
 {
 }
